sys/hashes/sha256: extract common code
